[FIX] web_editor: unlink only the selected link
Before this commit when a button containing a text then an icon was
unlinked in Chrome, all further buttons were also unlinked.

After this commit "contenteditable" is only activated on the parent node
of the link to prevent Chrome from selecting other nodes.
Also, Firefox replaces the `<a>` by a `<span>` during `unlink` - thus
keeping the `style` and `class` of the former attribute - therefore
this commit also removes those attributes before calling the command.

Steps to reproduce:
- drop Text+Image snippet twice
- inside first button type /image and insert a pictogram after the text
- inside first button type /image and insert a pictogram before or after the text
- click on the text of the first button => the link popover appears
- click on the unlink button of the popover
=> both buttons become flat texts in Chrome
=> nothing happens visually in Firefox, but the `<a>` actually gets
turned into a `<span>`

task-2778912